requirements.txt
flask
mysql-connector-python
qrcode[pil]
opencv-python
numpy